Use rich text editor in translations
