+++ src/storybook/stories/pages/PageDetailsPage.tsx
@@ -1,9 +1,9 @@
+import { PageData } from "@saleor/pages/components/PageDetailsPage/form";
 import { PageErrorCode } from "@saleor/types/globalTypes";
 import { storiesOf } from "@storybook/react";
 import React from "react";
 
 import PageDetailsPage, {
-  PageDetailsPageFormData,
   PageDetailsPageProps
 } from "../../../pages/components/PageDetailsPage";
 import { page } from "../../../pages/fixtures";
@@ -36,7 +36,7 @@ storiesOf("Views / Pages / Page details", module)
         "isPublished",
         "seoDescription",
         "seoTitle"
-      ] as Array<keyof PageDetailsPageFormData>).map(field => ({
+      ] as Array<keyof PageData>).map(field => ({
         __typename: "PageError",
         code: PageErrorCode.INVALID,
         field
