+++ src/translations/types.ts
@@ -1,3 +1,4 @@
+import { OutputData } from "@editorjs/editorjs";
 import { ConfirmButtonTransitionState } from "@saleor/components/ConfirmButton";
 import { ShopInfo_shop_languages } from "@saleor/components/Shop/types/ShopInfo";
 
@@ -11,5 +12,5 @@ export interface TranslationsEntitiesPageProps {
   onEdit: (field: string) => void;
   onDiscard: () => void;
   onLanguageChange: (lang: string) => void;
-  onSubmit: (field: string, data: string) => void;
+  onSubmit: (field: string, data: string | OutputData) => void;
 }
